+++ docs/changelog.rst
@@ -10,6 +10,8 @@ Changelog
   per-disc. For example, if track 24 on a release is the first track on the
   second disc, then it is not penalized for having its track number set to 1
   instead of 24.
+* The autotagger now also tolerates tracks whose track artists tags are set
+  to "Various Artists".
 * Fix a bug in the ``rewrite`` plugin that broke the use of multiple rules for
   a single field.